fix(upload): fixed unit tests
