+++ zboxcore/sdk/allocation_test.go
@@ -547,201 +547,6 @@ func TestAllocation_CreateDir(t *testing.T) {
 	require.NoErrorf(err, "Unexpected error %v", err)
 }
 
-func TestAllocation_RepairFile(t *testing.T) {
-	const (
-		mockFileRefName = "mock file ref name"
-		mockLocalPath   = "1.txt"
-		mockActualHash  = "75a919d23622c29ade8096ed1add6606ec970579459178db3a7d1d0ff8df92d3"
-		mockChunkHash   = "a6fb1cb61c9a3b8709242de28e44fb0b4de3753995396ae1d21ca9d4e956e9e2"
-	)
-
-	var mockClient = mocks.HttpClient{}
-	zboxutil.Client = &mockClient
-
-	client := zclient.GetClient()
-	client.Wallet = &zcncrypto.Wallet{
-		ClientID:  mockClientId,
-		ClientKey: mockClientKey,
-	}
-
-	setupHttpResponses := func(t *testing.T, testName string, numBlobbers, numCorrect int) {
-		require.True(t, numBlobbers >= numCorrect)
-		for i := 0; i < numBlobbers; i++ {
-			var hash string
-			if i < numCorrect {
-				hash = mockActualHash
-			}
-			frName := mockFileRefName + strconv.Itoa(i)
-			url := "TestAllocation_RepairFile" + testName + mockBlobberUrl + strconv.Itoa(i) + "/v1/file/meta"
-			mockClient.On("Do", mock.MatchedBy(func(req *http.Request) bool {
-				return strings.HasPrefix(req.URL.Path, url)
-			})).Return(&http.Response{
-				StatusCode: http.StatusOK,
-				Body: func(fileRefName, hash string) io.ReadCloser {
-					jsonFR, err := json.Marshal(&fileref.FileRef{
-						ActualFileHash: hash,
-						Ref: fileref.Ref{
-							Name: fileRefName,
-						},
-					})
-					require.NoError(t, err)
-					return ioutil.NopCloser(bytes.NewReader([]byte(jsonFR)))
-				}(frName, hash),
-			}, nil)
-		}
-	}
-
-	setupHttpResponsesWithUpload := func(t *testing.T, testName string, numBlobbers, numCorrect int) {
-		require.True(t, numBlobbers >= numCorrect)
-		for i := 0; i < numBlobbers; i++ {
-			var hash string
-			if i < numCorrect {
-				hash = mockActualHash
-			}
-
-			frName := mockFileRefName + strconv.Itoa(i)
-			httpResponse := &http.Response{
-				StatusCode: http.StatusOK,
-				Body: func(fileRefName, hash string) io.ReadCloser {
-					jsonFR, err := json.Marshal(&fileref.FileRef{
-						ActualFileHash: hash,
-						Ref: fileref.Ref{
-							Name: fileRefName,
-						},
-					})
-					require.NoError(t, err)
-					return ioutil.NopCloser(bytes.NewReader([]byte(jsonFR)))
-				}(frName, hash),
-			}
-
-			urlMeta := "TestAllocation_RepairFile" + testName + mockBlobberUrl + strconv.Itoa(i) + "/v1/file/meta"
-			mockClient.On("Do", mock.MatchedBy(func(req *http.Request) bool {
-				return strings.HasPrefix(req.URL.Path, urlMeta)
-			})).Return(httpResponse, nil)
-
-			urlUpload := "TestAllocation_RepairFile" + testName + mockBlobberUrl + strconv.Itoa(i) + "/v1/file/upload"
-			mockClient.On("Do", mock.MatchedBy(func(req *http.Request) bool {
-				return strings.HasPrefix(req.URL.Path, urlUpload)
-			})).Return(&http.Response{
-				StatusCode: http.StatusOK,
-				Body: func(fileRefName, hash string) io.ReadCloser {
-					jsonFR, err := json.Marshal(&UploadResult{
-						Filename: mockLocalPath,
-						Hash:     mockChunkHash,
-					})
-					require.NoError(t, err)
-					return ioutil.NopCloser(bytes.NewReader([]byte(jsonFR)))
-				}(frName, hash),
-			}, nil)
-
-			urlFilePath := "TestAllocation_RepairFile" + testName + mockBlobberUrl + strconv.Itoa(i) + "/v1/file/referencepath"
-			mockClient.On("Do", mock.MatchedBy(func(req *http.Request) bool {
-				return strings.HasPrefix(req.URL.Path, urlFilePath)
-			})).Return(&http.Response{
-				StatusCode: http.StatusOK,
-				Body: func(fileRefName, hash string) io.ReadCloser {
-					jsonFR, err := json.Marshal(&ReferencePathResult{
-						ReferencePath: &fileref.ReferencePath{
-							Meta: map[string]interface{}{
-								"type": "d",
-							},
-						},
-						LatestWM: nil,
-					})
-					require.NoError(t, err)
-					return ioutil.NopCloser(bytes.NewReader([]byte(jsonFR)))
-				}(frName, hash),
-			}, nil)
-
-			urlCommit := "TestAllocation_RepairFile" + testName + mockBlobberUrl + strconv.Itoa(i) + "/v1/connection/commit"
-			mockClient.On("Do", mock.MatchedBy(func(req *http.Request) bool {
-				return strings.HasPrefix(req.URL.Path, urlCommit)
-			})).Return(&http.Response{
-				StatusCode: http.StatusOK,
-				Body: func(fileRefName, hash string) io.ReadCloser {
-					jsonFR, err := json.Marshal(&ReferencePathResult{})
-					require.NoError(t, err)
-					return ioutil.NopCloser(bytes.NewReader([]byte(jsonFR)))
-				}(frName, hash),
-			}, nil)
-		}
-	}
-
-	type parameters struct {
-		localPath  string
-		remotePath string
-		status     StatusCallback
-	}
-	tests := []struct {
-		name        string
-		parameters  parameters
-		numBlobbers int
-		numCorrect  int
-		setup       func(*testing.T, string, int, int)
-		wantErr     bool
-		errMsg      string
-	}{
-		{
-			name: "Test_Repair_Not_Required_Failed",
-			parameters: parameters{
-				localPath:  mockLocalPath,
-				remotePath: "/",
-			},
-			numBlobbers: 4,
-			numCorrect:  4,
-			setup:       setupHttpResponses,
-			wantErr:     true,
-			errMsg:      "chunk_upload: Repair not required",
-		},
-		{
-			name: "Test_Repair_Required_Success",
-			parameters: parameters{
-				localPath:  mockLocalPath,
-				remotePath: "/",
-			},
-			numBlobbers: 6,
-			numCorrect:  5,
-			setup:       setupHttpResponsesWithUpload,
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			require := require.New(t)
-			if teardown := setupMockFile(t, mockLocalPath); teardown != nil {
-				defer teardown(t)
-			}
-			a := &Allocation{
-				ParityShards: tt.numBlobbers / 2,
-				DataShards:   tt.numBlobbers / 2,
-			}
-			a.uploadChan = make(chan *UploadRequest, 10)
-			a.downloadChan = make(chan *DownloadRequest, 10)
-			a.repairChan = make(chan *RepairRequest, 1)
-			a.ctx, a.ctxCancelF = context.WithCancel(context.Background())
-			a.uploadProgressMap = make(map[string]*UploadRequest)
-			a.downloadProgressMap = make(map[string]*DownloadRequest)
-			a.mutex = &sync.Mutex{}
-			a.initialized = true
-			sdkInitialized = true
-			setupMockAllocation(t, a)
-			for i := 0; i < tt.numBlobbers; i++ {
-				a.Blobbers = append(a.Blobbers, &blockchain.StorageNode{
-					Baseurl: "TestAllocation_RepairFile" + tt.name + mockBlobberUrl + strconv.Itoa(i),
-				})
-			}
-			tt.setup(t, tt.name, tt.numBlobbers, tt.numCorrect)
-			err := a.RepairFile(tt.parameters.localPath, tt.parameters.remotePath, tt.parameters.status)
-			require.EqualValues(tt.wantErr, err != nil)
-			if err != nil {
-				require.EqualValues(tt.errMsg, errors.Top(err))
-				return
-			}
-			require.NoErrorf(err, "Unexpected error %v", err)
-		})
-	}
-}
-
 func TestAllocation_RepairRequired(t *testing.T) {
 	const (
 		mockActualHash = "4041e3eeb170751544a47af4e4f9d374e76cee1d"
